move soffice serialization to conversion
